updated profile, stageView, sharing, and meeting
