+++ change/@microsoft-teams-js-262cb870-a488-4819-8561-7f3c35051e33.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Made the `meeting`, `profile`, `sharing`, and `stageView` files treeshakable",
+  "packageName": "@microsoft/teams-js",
+  "email": "109628470+noahdarveau-MSFT@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
