+++ packages/teams-js/src/public/index.ts
@@ -66,17 +66,17 @@ export * as menus from './menus';
 export * as media from './media';
 export * as secondaryBrowser from './secondaryBrowser';
 export * as location from './location';
-export { meeting } from './meeting';
+export * as meeting from './meeting/meeting';
 export * as monetization from './monetization';
 export * as calendar from './calendar';
 export * as mail from './mail';
 export * as teamsCore from './teamsAPIs';
 export * as people from './people';
-export { profile } from './profile';
+export * as profile from './profile';
 export { videoEffects } from './videoEffects';
 export * as search from './search';
-export { sharing } from './sharing';
-export { stageView } from './stageView';
+export * as sharing from './sharing/sharing';
+export * as stageView from './stageView/stageView';
 export { version } from './version';
 export { visualMedia } from './visualMedia';
 export { webStorage } from './webStorage';
