+++ packages/teams-js/test/public/profile.spec.ts
@@ -2,8 +2,8 @@ import { errorLibraryNotInitialized } from '../../src/internal/constants';
 import * as app from '../../src/public/app/app';
 import { FrameContexts } from '../../src/public/constants';
 import { ErrorCode } from '../../src/public/interfaces';
-import { ShowProfileRequestInternal } from '../../src/public/profile';
-import { profile } from '../../src/public/profile';
+import { ShowProfileRequestInternal } from '../../src/internal/profileUtil';
+import * as profile from '../../src/public/profile';
 import { Utils } from '../utils';
 
 /* eslint-disable */
